adicionando a funcionalidade de editar produto e melhoria nas paginas

+++ routes/web.php
@@ -11,8 +11,9 @@ Route::prefix('produtos')->group(function () {
     Route::match(['get', 'post'], '/criar', [StoreController::class, 'create'])->middleware('can:admin')->name('produtos-criar');
     Route::match(['get', 'post'], '/list', [StoreController::class, 'list'])->middleware('can:admin')->name('produtos-list');
     Route::match(['get', 'post'], '/{id}', [StoreController::class, 'show'])->name('produtos-id');
-    Route::match(['get', 'post'], '/{id}', [StoreController::class, 'edit'])->name('produtos-edit');
-    Route::delete('/{id}', [StoreController::class, 'destroy'])->name('produtos-delete-id');
+    Route::match(['get', 'post'], '/editar/{id}', [StoreController::class, 'edit'])->name('produtos-edit');
+    Route::delete('/{id}', [StoreController::class, 'destroy'])->middleware('can:admin')->name('produtos-delete-id');
+    Route::put('/update/{id}', [StoreController::class, 'update'])->middleware('can:admin')->name('produtos-uodate');
 });
 
 Route::match(['get', 'post'], '/categoria', [StoreController::class, 'category'])->name('categoria');
